+++ api/src/main/java/org/openmrs/module/labonfhir/LabOnFhirConfig.java
@@ -68,6 +68,12 @@ public class LabOnFhirConfig implements ApplicationContextAware {
 	
 	public static final String GP_FILTER_ORDER_BY_TEST_UUIDS = "labonfhir.filterOrderBytestUuids";
 	
+	public static final String GP_REST_HOOK_USERNAME = "labonfhir.restHook.username";
+	
+	public static final String GP_REST_HOOK_PASSWORD = "labonfhir.restHook.password";
+	
+	public static final String GP_REST_HOOK_BASE_ENDPOINT = "labonfhir.restHook.baseEndpoint";
+	
 	public enum AuthType {
 		SSL,
 		BASIC
@@ -111,7 +117,7 @@ public class LabOnFhirConfig implements ApplicationContextAware {
 	}
 	
 	public String getLisUrl() {
-		//return GP_OPENELIS_URL
+		// return GP_OPENELIS_URL
 		String url = administrationService.getGlobalProperty(GP_LIS_URL);
 		
 		if (StringUtils.isBlank(url)) {
