+++ api/src/main/java/org/openmrs/module/labonfhir/api/service/FetchTaskUpdatesService.java
@@ -1,0 +1,13 @@
+package org.openmrs.module.labonfhir.api.service;
+
+import java.util.UUID;
+
+import javax.annotation.Nullable;
+
+import org.openmrs.api.OpenmrsService;
+
+public interface FetchTaskUpdatesService extends OpenmrsService {
+	
+	public void fetchTaskUpdates(@Nullable UUID uuid);
+	
+}
